added more php code

+++ lib/php/user.php
@@ -181,7 +181,7 @@ public function setZip($newZip) {
 
     public static function getAllUsers(PDO $pdo) {
         //create the query template
-        $query = "SELECT userID, name, email,phone FROM user";
+        $query = "SELECT userID, name, email, phone FROM user";
         $statement = $pdo->prepare($query);
         // execute
         $statement->execute();
@@ -193,7 +193,7 @@ public function setZip($newZip) {
         try {
             if($row !== false) {
                 $user = new user($row["userID"], $row["name"], $row["email"], $row["phone"]);
-                $user[$user->key()] = $user;
+                $userID[$userID->key()] = $user;
                 $user->next();
                 }
             } catch(Exception $exception) {
@@ -211,26 +211,31 @@ public function setZip($newZip) {
         return ($fields);
     }
 
-    public static function getUserByUserID(PDO $pdo, $userID) {
-        $userID = filter_var($userID, FILTER_VALIDATE_INT);
+    public static function getUserByUserID(PDO $pdo, $user) {
+
+        $userID = filter_var($user, FILTER_VALIDATE_INT);
+
         if($userID === false) {
             throw(new PDOException(""));
+        }
+        if($userID <= 0) {
+            throw(new PDOException("userID id is not positive"));
         }
         // create query template
         $query = "SELECT userID, email, name, phone FROM user WHERE userID = :userID";
         $statement = $pdo->prepare($query);
         // bind the bulletinid to the place holder in the template
-        $parameters = array("userID" => $userID);
+        $parameters = array("user" => $user);
         $statement->execute($parameters);
 
-        $userID= null;
+        $user= null;
         $statement->setFetchMode(PDO::FETCH_ASSOC);
         $row = $statement->fetch();
 
         // grab the bulletin from mySQL
         try {
             if($row !== false) {
-                $userID = new user ($row["userId"], $row["name"], $row["email"], $row["phone"]);
+                $user = new user ($row["userId"], $row["name"], $row["email"], $row["phone"]);
                 $userID[$userID->key()] = $user;
                 $userID->next();
             }
@@ -238,7 +243,7 @@ public function setZip($newZip) {
             // if the row couldn't be converted, rethrow it
             throw(new PDOException($exception->getMessage(), 0, $exception));
         }
-        return $userID;
+        return $user;
     }
 
 
@@ -247,10 +252,10 @@ public function setZip($newZip) {
     // make sure user doesn't already exist
     if ($this->userID !== null) {
         throw (new PDOException("existing user"));
-
+    }
         //create query template
         $query
-            = "INSERT INTO user(name, email, phone, zip, salt, hash)" .
+            = "INSERT INTO 'user'(name, email, phone, zip, salt, hash)" .
                 "VALUES ( :name, :email, :hash, :phone, :salt, :zip)";
         $statement = $pdo->prepare($query);
         // bind the variables to the place holders in the template
@@ -259,22 +264,28 @@ public function setZip($newZip) {
             "salt" => $this->salt, "zip" => $this->zip);
         $statement->execute($parameters);
         //update null userId with what mySQL just gave us
-        $this->userID = intval($pdo->lastInsertID);
+        $this->userID = intval($pdo->lastInsertId());
 
     }
-}
+
 
     public function delete(PDO $pdo) {
+
+        if ($this->userID === null){
+            throw(new PDOException("unable to delete a project that does not exist"));
+        }
+
         //create query template
         $query = "DELETE FROM 'user' WHERE userID = :userID";
         $statement = $pdo->prepare($query);
+
         $parameters = array("userID" => $this->userID);
         $statement->execute($parameters);
     }
 
     public function update(PDO $pdo) {
         // create query template
-        $query = "UPDATE user SET userID = :userID, name = :name, email = :email, phone = :phone, zip = :zip WHERE userID = :userID, name = :name, email = :email, phone = :phone, zip = :zip";
+        $query = "UPDATE 'user' SET name = :name, email = :email, phone = :phone, zip = :zip WHERE userID = :userID";
         $statement = $pdo->prepare($query);
         // bind the member variables
         $parameters = array("userID" => $this->userID, "name" => $this->name, "email" => $this->email, "phone"=> $this->phone, "zip" => $this->zip);
